Refactored MagicLinks and added account deletion
In order to add account deletion by magic link as a feature, I decided
to go ahead and completely refactor magic_links to allow for the easy
inclusion of additional magic_link paths. The system can still stand to be
refactored further, but for now it is in working order, and allows for
new magic link paths to be easily added. Closes #115.