Handle Disabled Slugs
Disabled slugs are now handled, and disabling of slugs now robust.
A function was created in UserController for generating new slugs.
It is currently incomplete, as if you disable the generated slug
too many times consecutively, it can cause crashes on Repo.update.
This is because it generates the slug based on how many users with
the same name as the user. If you're generating new slugs without
creating new users with the same name, it will generate the same
slug every time. A module for generating new slugs might be
necessary. Closes #91.