Implemented Facebook Login
The login screen now sports a facebook login link. Users are created
in our system after they login from facebook. Note that currently,
the only data from facebook added is first and last name. This will
be expanded on later, but there is an issue with the facebook api
not returning the requested fields for all users. The only fields I
could garaunteed get for now are first and last name. The user can
still add fields just fine to their account, and use facebook to
log in, so having an email is not necessesary, though if they add an
email, they can use it to log in. Closes #99.