Added user social profile information
When the user adds social media accounts, they will automatically be
included in the application/id+json of their User#show. However the
url field is currently not used, as we do not have a way for the
user to specify their "personal" website. We might consider adding
this later on.
Closes #98.