Added read-only restful API
A read-only API has been implemented for the following resources:
User, address, email, phone_number, skill, user_skill,
social_media_account, and work_experience. The rest is soon to come.
Much of the code in the controllers for the api resources is
commented out. This is because those are the write functions of the
API, which is not yet ready. They will be uncommented when that
portion of the api is implemented.
Progress towards #142.

The read-only api is not yet complete.