Style progress, restricted search_terms, SMTP
Comitting some styling progress. Changes in this category include
a few helper functions and additional assigns in user#show. Main
change is a new layout and template for user#show.

Restricted search_terms, users cannot directly modify their own
search terms. They are regenerated each time a user edits their
information, ensuring they are always correct.

SMTP server integration is enabled. Because of this, you will not
be able to log in without something to recieve the email with the
magic link, be it a dev smtp similator, or a functional smtp server
to send you the email.

Closes #167, progress twards #100.