Update readme with secret.exs instructions
With the addition of a secret config, it is now required to have the secret config files or else the application will not compile. They are not included in the source files for security reasons. To reflect this change, the readme now instructs users to create it as part of the setup instructions.