Gravatar downloading and locale storage.
The user's locale is now stored in the locale field. It is
automatically set when the user is created, and can be edited
in the user#edit form. Closes #173

Gravatar is now downloaded when the user is created and uploaded
via arc-ecto. It is temporarily stored in the local machines temp
directory. This needs testing on multiple machines to ensure it
is platform independant. In a future version arc is adding support
for storing binary data, which will hopefully remove the need for
me storing the file in a temp location before uploading it.
Closes #172.

Added a new plug to make progress towards closing #174. It is a
plug which drops session when a user is not logged in. It is not
complete, but does achieve the desired result of not storing cookies.
It has the undesired result of causing crsf errors whenever the user
tries to submit forms. Further development is needed. To test,
uncomment it in the router.