Unlocked and upgraded /skills route
Skills allows for the viewing of skill#show pages. These pages
feature lists of users that have the skill that are either
top in endorsements, or connected to the current user. Also
includes a button to add the skill if the current_user
does not have the skill. Closes #478. Closes #479.