Changed user_skills to use pill css
Some things to note: I jerry rigged delete buttons for the the
user to be able to delete user_skills, as the pills did not
have that functionality built in. Additionally because the
skills are made to use <a> tags I had to improvive to get
them to work with the <form><button> tags we use.

Closes #510