Added rudimentary search to tags.
Most of the logic used is reused skill code as a more appropriate
system would just be replaced by the more robust search system
to be implemented in the near future, but if the search system
were to be removed, it would provide a bad experience for users.

Also fixes a bug that prevented the deletion of user_tags.
Closes #528